api/requirements.txt
Flask
numpy
pytest
flake8
PyMySQL
pytest-cov
python-magic
opencv-python